Refactor and add tests for search_coauthors()
Separated the part of search_authors which actually requests the data
from the Ajax handling and sanitization functions. Added tests for the
functional part of this feature.

Defined a comparator function for excluding existing authors on a post
from the array returned by search_coauthors() - because the authors in
the existing array have an additional attribute ("author_role") added, a
simple array_diff won't work.