Fix bug in `get_coauthors` query
`get_coauthors` shouldn't return byline authors if another role is
requested. It was returning any authors set on a post with the coauthor
taxonomy, even if a different author_role was requested.

Also simplified the logic a bit ans structured the tests of it a little
better so I can work towards doing a real refactor of that area.