Avoid warning when opening new post
The sortable template references the "author_role" property of the coauthor.
In the case of a new post which hasn't been saved yet, this property won't
exist (because the author object is straight from Co-Authors Plus, which
doesn't have that field.)

Make sure to set that field on any Authors on a post, to avoid this PHP
warning.