Always sort byline authors before other roles
Because there isn't a defined sort order between coauthors who are set
as taxonomy terms on a post (as in Co-Authors Plus), and coauthors who
are set in meta fields (as in this plugin), there isn't going to be
consistant behavior if we allow sorting the two together.

This commit changes the `get_coauthors()` response so that it always
returns the byline authors before the other roles. It does not yet
prevent the UI from sorting another role higher - its possible to do
that, but that order will revert on save.