Escape and sanitize all the things!
- Escape all strings from database objects or translations which will be
  printed.
- Sanitize all user input before saving

In some places this might be overkill: I tried to optimize for clarity
in reading so that a quick glance would show what filters a given
variable is going through. Sometimes this means a given value may be
passed through the same function multiple times; when its received from
input and then later in the funciton it's actually sent to.

Fixes #2.