Add the ability to specify a list of users to exclude in search
The only filtering available in the search_coauthors() function was just
to exclude all authors on the current post. This isn't fine grained
enough for the UI, which needs to know which authors are currently
added, whether or not they've been saved yet. I added an optional third
argument to the function where you can specify an array of nicenames to
exclude.