Change search_coauthor to use wp_parse_args format
Eventually this will need to handle pagination, etc. I don't think
that's necessary for initial plugin release, but for sites with huge
numbers of authors it'll be necessary sooner or later.

Adding any additional positional arguments to this function will get
messy real quickly; better to switch to the named argument style used in
core now.